+++ src/config/structs.rs
@@ -52,11 +52,9 @@ pub enum ModLoader {
     Fabric,
     Forge,
 }
-
 #[derive(thiserror::Error, Debug, PartialEq)]
 #[error("The given string is not a mod loader")]
 pub struct ModLoaderParseError {}
-
 impl TryFrom<&String> for ModLoader {
     type Error = ModLoaderParseError;
     fn try_from(from: &String) -> Result<Self, Self::Error> {